+++ visualization/app/codeCharta/state/settingsService/settings.service.ts
@@ -11,6 +11,7 @@ import {
 	BlacklistSubscriber,
 	ColorMetricSubscriber,
 	DistributionMetricSubscriber,
+	EdgeMetricSubscriber,
 	HeightMetricSubscriber,
 	SearchPatternSubscriber,
 	SettingsEvents,
@@ -31,6 +32,20 @@ export class SettingsService implements FileStateServiceSubscriber {
 			this.debounceBroadcast[SettingsEvents[key]] = _.debounce((eventName: string, data: any) => {
 				if (eventName == SettingsEvents.SETTINGS_CHANGED_EVENT) {
 					this.update = {}
+				} else if (this.update.dynamicSettings && eventName == SettingsEvents.EDGE_METRIC_CHANGED_EVENT) {
+					delete this.update.dynamicSettings.edgeMetric
+				} else if (this.update.dynamicSettings && eventName == SettingsEvents.AREA_METRIC_CHANGED_EVENT) {
+					delete this.update.dynamicSettings.areaMetric
+				} else if (this.update.dynamicSettings && eventName == SettingsEvents.HEIGHT_METRIC_CHANGED_EVENT) {
+					delete this.update.dynamicSettings.heightMetric
+				} else if (this.update.dynamicSettings && eventName == SettingsEvents.COLOR_METRIC_CHANGED_EVENT) {
+					delete this.update.dynamicSettings.colorMetric
+				} else if (this.update.fileSettings && eventName == SettingsEvents.BLACKLIST_CHANGED_EVENT) {
+					delete this.update.fileSettings.blacklist
+				} else if (this.update.dynamicSettings && eventName == SettingsEvents.DISTRIBUTION_METRIC_CHANGED_EVENT) {
+					delete this.update.dynamicSettings.distributionMetric
+				} else if (this.update.dynamicSettings && eventName == SettingsEvents.SEARCH_PATTERN_CHANGED_EVENT) {
+					delete this.update.dynamicSettings.searchPattern
 				}
 				this.$rootScope.$broadcast(eventName, data)
 			}, SettingsService.DEBOUNCE_TIME)
@@ -72,6 +87,10 @@ export class SettingsService implements FileStateServiceSubscriber {
 					this.notifyColorMetricSubscribers()
 				}
 
+				if (this.update.dynamicSettings.edgeMetric) {
+					this.notifyEdgeMetricSubscribers()
+				}
+
 				if (this.update.dynamicSettings.distributionMetric) {
 					this.notifyDistributionMetricSubscribers()
 				}
@@ -98,7 +117,9 @@ export class SettingsService implements FileStateServiceSubscriber {
 			flat: "#AAAAAA",
 			lightGrey: "#DDDDDD",
 			angularGreen: "#00BFA5",
-			markingColors: ["#FF1D8E", "#1d8eff", "#1DFFFF", "#8eff1d", "#8e1dff", "#FFFF1D"]
+			markingColors: ["#FF1D8E", "#1d8eff", "#1DFFFF", "#8eff1d", "#8e1dff", "#FFFF1D"],
+			incomingEdge: "#00ffff",
+			outgoingEdge: "#ff00ff"
 		}
 
 		const scaling: Vector3 = new Vector3(1, 1, 1)
@@ -117,6 +138,7 @@ export class SettingsService implements FileStateServiceSubscriber {
 				heightMetric: null,
 				colorMetric: null,
 				distributionMetric: null,
+				edgeMetric: null,
 				focusedNodePath: "",
 				searchedNodePaths: [],
 				searchPattern: "",
@@ -125,9 +147,10 @@ export class SettingsService implements FileStateServiceSubscriber {
 			},
 			appSettings: {
 				amountOfTopLabels: 1,
+				amountOfEdgePreviews: 1,
+				edgeHeight: 4,
 				scaling: scaling,
 				camera: camera,
-				enableEdgeArrows: true,
 				hideFlatBuildings: true,
 				maximizeDetailPanel: false,
 				invertColorRange: false,
@@ -137,7 +160,8 @@ export class SettingsService implements FileStateServiceSubscriber {
 				isWhiteBackground: false,
 				whiteColorBuildings: false,
 				mapColors: mapColors,
-				isPresentationMode: false
+				isPresentationMode: false,
+				showOnlyBuildingsWithEdges: false
 			},
 			treeMapSettings: {
 				mapSize: 500
@@ -234,6 +258,10 @@ export class SettingsService implements FileStateServiceSubscriber {
 		this.notify(SettingsEvents.COLOR_METRIC_CHANGED_EVENT, { colorMetric: this.settings.dynamicSettings.colorMetric })
 	}
 
+	private notifyEdgeMetricSubscribers() {
+		this.notify(SettingsEvents.EDGE_METRIC_CHANGED_EVENT, { edgeMetric: this.settings.dynamicSettings.edgeMetric })
+	}
+
 	private notifyDistributionMetricSubscribers() {
 		this.notify(SettingsEvents.DISTRIBUTION_METRIC_CHANGED_EVENT, {
 			distributionMetric: this.settings.dynamicSettings.distributionMetric
@@ -279,6 +307,12 @@ export class SettingsService implements FileStateServiceSubscriber {
 	public static subscribeToColorMetric($rootScope: IRootScopeService, subscriber: ColorMetricSubscriber) {
 		$rootScope.$on(SettingsEvents.COLOR_METRIC_CHANGED_EVENT, (event, data) => {
 			subscriber.onColorMetricChanged(data.colorMetric)
+		})
+	}
+
+	public static subscribeToEdgeMetric($rootScope: IRootScopeService, subscriber: EdgeMetricSubscriber) {
+		$rootScope.$on(SettingsEvents.EDGE_METRIC_CHANGED_EVENT, (event, data) => {
+			subscriber.onEdgeMetricChanged(data.edgeMetric)
 		})
 	}