+++ visualization/app/codeCharta/ui/edgeSettingsPanel/edgeSettingsPanel.module.ts
@@ -1,0 +1,7 @@
+import "../../state/state.module"
+import angular from "angular"
+import { edgeSettingsPanelComponent } from "./edgeSettingsPanel.component"
+
+angular
+	.module("app.codeCharta.ui.edgeSettingsPanel", ["app.codeCharta.state"])
+	.component(edgeSettingsPanelComponent.selector, edgeSettingsPanelComponent)